+++ bench/irmin-pack/tree.ml
@@ -195,7 +195,10 @@ module Make_store_layered (Conf : Irmin_pack.Conf.S) = struct
   include Store
 end
 
-module Make_basic (Maker : Irmin_pack.Maker) (Conf : Irmin_pack.Conf.S) = struct
+module Make_basic (Maker : functor (_ : Irmin_pack.Conf.S) ->
+  Irmin_pack.Maker)
+(Conf : Irmin_pack.Conf.S) =
+struct
   type store_config = config
 
   module Store = struct