+++ src/irmin-pack/ext.ml
@@ -20,21 +20,21 @@ module IO = IO.Unix
 module Maker (V : Version.S) (Config : Conf.S) = struct
   type endpoint = unit
 
-  include Irmin.Key.Store_spec.Hash_keyed
+  include Pack_key.Store_spec
 
   module Make (Schema : Irmin.Schema.Extended) = struct
     open struct
-      module H = Schema.Hash
       module P = Schema.Path
       module M = Schema.Metadata
       module C = Schema.Contents
       module B = Schema.Branch
     end
 
+    module H = Schema.Hash
     module Index = Pack_index.Make (H)
     module Pack = Pack_store.Maker (V) (Index) (H)
     module Dict = Pack_dict.Make (V)
-    module XKey = Irmin.Key.Of_hash (H)
+    module XKey = Pack_key.Make (H)
 
     module X = struct
       module Hash = H
@@ -43,16 +43,16 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
       [@@deriving irmin]
 
       module Contents = struct
-        module Pack_value = Pack_value.Of_contents (H) (C)
+        module Pack_value = Pack_value.Of_contents (H) (XKey) (C)
         module CA = Pack.Make (Pack_value)
-        include Irmin.Contents.Store (CA) (H) (C)
+        include Irmin.Contents.Store_indexable (CA) (H) (C)
       end
 
       module Node = struct
         module Value = Schema.Node (XKey) (XKey)
 
         module CA = struct
-          module Inter = Inode.Make_internal (Config) (H) (Value)
+          module Inter = Inode.Make_internal (Config) (H) (XKey) (Value)
           include Inode.Make_persistent (H) (Value) (Inter) (Pack)
         end
 
@@ -75,29 +75,22 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
         end
 
         module Pack_value =
-          Pack_value.Of_commit
-            (H)
+          Pack_value.Of_commit (H) (XKey)
             (struct
               module Info = Schema.Info
               include Value
             end)
 
         module CA = Pack.Make (Pack_value)
-        include Irmin.Commit.Store (Schema.Info) (Node) (CA) (H) (Value)
+
+        include
+          Irmin.Commit.Generic_key.Store (Schema.Info) (Node) (CA) (H) (Value)
       end
 
       module Branch = struct
         module Key = B
-
-        module Val = struct
-          include H
-          include Commit.Key
-        end
-
-        module AW =
-          Atomic_write.Make_persistent (V) (Key)
-            (Atomic_write.Value.Of_hash (Val))
-
+        module Val = XKey
+        module AW = Atomic_write.Make_persistent (V) (Key) (Val)
         include Atomic_write.Closeable (AW)
 
         let v ?fresh ?readonly path =
@@ -105,7 +98,7 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
       end
 
       module Slice = Irmin.Backend.Slice.Make (Contents) (Node) (Commit)
-      module Remote = Irmin.Backend.Remote.None (H) (B)
+      module Remote = Irmin.Backend.Remote.None (Commit.Key) (B)
 
       module Repo = struct
         type t = {
@@ -229,7 +222,7 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
       let* heads =
         match heads with None -> Repo.heads t | Some m -> Lwt.return m
       in
-      let hashes = List.map (fun x -> `Commit (Commit.hash x)) heads in
+      let hashes = List.map (fun x -> `Commit (Commit.key x)) heads in
       let+ () =
         Repo.iter ~cache_size:1_000_000 ~min:[] ~max:hashes ~node ~commit t
       in
@@ -245,7 +238,7 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
           !errors
 
     module Stats = struct
-      let pp_hash = Irmin.Type.pp Hash.t
+      let pp_key = Irmin.Type.pp XKey.t
 
       let traverse_inodes ~dump_blob_paths_to commit repo =
         let module Stats = Checks.Stats (struct
@@ -258,12 +251,19 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
         let t = Stats.v () in
         let pred_node repo k =
           X.Node.find (X.Repo.node_t repo) k >|= function
-          | None -> Fmt.failwith "hash %a not found" pp_hash k
+          | None -> Fmt.failwith "key %a not found" pp_key k
           | Some v ->
               let width = X.Node.Val.length v in
               let nb_children = X.Node.CA.Val.nb_children v in
               let preds = X.Node.CA.Val.pred v in
-              Stats.visit_node t k ~width ~nb_children preds;
+              let () =
+                preds
+                |> List.map (function
+                     | s, `Contents h -> (s, `Contents (XKey.to_hash h))
+                     | s, `Inode h -> (s, `Inode (XKey.to_hash h))
+                     | s, `Node h -> (s, `Node (XKey.to_hash h)))
+                |> Stats.visit_node t (XKey.to_hash k) ~width ~nb_children
+              in
               List.rev_map
                 (function
                   | s, `Inode x ->
@@ -279,11 +279,11 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
           | None -> []
           | Some c ->
               let node = X.Commit.Val.node c in
-              Stats.visit_commit t node;
+              Stats.visit_commit t (XKey.to_hash node);
               [ `Node node ]
         in
         let pred_contents _repo k =
-          Stats.visit_contents t k;
+          Stats.visit_contents t (XKey.to_hash k);
           Lwt.return []
         in
         (* We want to discover all paths to a node, so we don't cache nodes
@@ -297,8 +297,8 @@ module Maker (V : Version.S) (Config : Conf.S) = struct
 
       let run ~dump_blob_paths_to ~commit repo =
         Printexc.record_backtrace true;
-        let hash = `Commit (Commit.hash commit) in
-        traverse_inodes ~dump_blob_paths_to hash repo
+        let key = `Commit (Commit.key commit) in
+        traverse_inodes ~dump_blob_paths_to key repo
     end
 
     let stats = Stats.run