+++ src/irmin-pack/mem/indexable.mli
@@ -13,5 +13,19 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
+open! Import
 
-module Maker (_ : Version.S) (_ : Conf.S) : S.Maker_persistent
+module Maker (K : Irmin.Hash.S) : sig
+  type key = K.t
+
+  module Make
+      (Val : Irmin_pack.Pack_value.S with type hash := K.t and type key := K.t) : sig
+    include
+      Irmin_pack.Indexable.S
+        with type hash = K.t
+         and type key = K.t
+         and type value = Val.t
+
+    val v : string -> read t Lwt.t
+  end
+end