+++ src/irmin-pack/pack_store_intf.ml
@@ -5,7 +5,7 @@ open! Import
     data blocks. The data file is indexed by hash via {!Pack_index.S}
     implementation. *)
 module type S = sig
-  include Content_addressable.S
+  include Indexable.S
 
   type index
 
@@ -33,18 +33,25 @@ module type S = sig
       are not removed. *)
 
   (** @inline *)
-  include S.Checkable with type 'a t := 'a t and type key := key
+  include S.Checkable with type 'a t := 'a t and type hash := hash
 end
 
 module type Maker = sig
-  type key
+  type hash
   type index
 
   (** Save multiple kind of values in the same pack file. Values will be
       distinguished using [V.magic], so they have to all be different. *)
 
-  module Make (V : Pack_value.S with type hash := key) :
-    S with type key = key and type value = V.t and type index = index
+  module Make
+      (V : Pack_value.Persistent
+             with type hash := hash
+              and type key := hash Pack_key.t) :
+    S
+      with type key = hash Pack_key.t
+       and type hash = hash
+       and type value = V.t
+       and type index := index
 end
 
 module type Sigs = sig
@@ -54,6 +61,6 @@ module type Sigs = sig
   module Maker
       (V : Version.S)
       (Index : Pack_index.S)
-      (K : Irmin.Hash.S with type t = Index.key) :
-    Maker with type key = K.t and type index = Index.t
+      (Hash : Irmin.Hash.S with type t = Index.key) :
+    Maker with type hash = Hash.t and type index := Index.t
 end