+++ test/irmin-pack/test_pack.ml
@@ -242,7 +242,7 @@ module Pack = struct
     let* () =
       Pack.batch t.pack (fun w ->
           Lwt_list.iter_s
-            (fun (k, v) -> Pack.unsafe_add w k v)
+            (fun (k, v) -> Pack.unsafe_add w k v >|= ignore)
             [ (h1, x1); (h2, x2); (h3, x3); (h4, x4) ])
     in
     let test t =
@@ -268,7 +268,10 @@ module Pack = struct
       let adds l =
         List.iter
           (fun (k, v) ->
-            Pack.unsafe_append ~ensure_unique:true ~overcommit:false w k v)
+            let (_ : Pack.hash) =
+              Pack.unsafe_append ~ensure_unique:true ~overcommit:false w k v
+            in
+            ())
           l
       in
       let x1 = "foo" in
@@ -304,7 +307,9 @@ module Pack = struct
     let* w1 = Pack.v ~fresh:true ~index (Context.fresh_name "pack") in
     let x1 = "foo" in
     let h1 = sha1 x1 in
-    Pack.unsafe_append ~ensure_unique:true ~overcommit:false w1 h1 x1;
+    let (_ : Pack.hash) =
+      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w1 h1 x1
+    in
     let* y1 = Pack.find w1 h1 >|= get in
     Alcotest.(check string) "x1" x1 y1;
     Index.close index;
@@ -316,7 +321,9 @@ module Pack = struct
     let w = t.pack in
     let x1 = "foo" in
     let h1 = sha1 x1 in
-    Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1;
+    let (_ : Pack.hash) =
+      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1
+    in
     Pack.flush w;
     Index.close t.index;
     Pack.close w >>= fun () ->
@@ -346,7 +353,9 @@ module Pack = struct
     let* () =
       Pack.batch w (fun w ->
           Lwt_list.iter_s
-            (fun (k, v) -> Pack.unsafe_add w k v)
+            (fun (k, v) ->
+              let+ (_ : Pack.hash) = Pack.unsafe_add w k v in
+              ())
             [ (h1, x1); (h2, x2) ])
     in
     Context.close t.index w >>= fun () ->
@@ -359,7 +368,9 @@ module Pack = struct
     (*open and close two packs *)
     let x3 = "toto" in
     let h3 = sha1 x3 in
-    Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h3 x3;
+    let (_ : Pack.hash) =
+      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h3 x3
+    in
     let* w2 = t.clone_pack ~readonly:false in
     Pack.close w >>= fun () ->
     let* y2 = Pack.find w2 h2 >|= get in
@@ -394,7 +405,9 @@ module Pack = struct
     let test w =
       let x1 = "foo" in
       let h1 = sha1 x1 in
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1;
+      let (_ : Pack.hash) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1
+      in
       Pack.sync r;
       let* y1 = Pack.find r h1 in
       Alcotest.(check (option string)) "sync before filter" None y1;
@@ -404,7 +417,9 @@ module Pack = struct
       Alcotest.(check (option string)) "sync after filter" (Some x1) y1;
       let x2 = "foo" in
       let h2 = sha1 x2 in
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h2 x2;
+      let (_ : Pack.hash) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h2 x2
+      in
       Index.flush t.index;
       let+ y2 = Pack.find r h2 in
       Alcotest.(check (option string)) "sync after flush" (Some x2) y2
@@ -422,7 +437,9 @@ module Pack = struct
     let test w =
       let x1 = "foo" in
       let h1 = sha1 x1 in
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1;
+      let (_ : Pack.hash) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h1 x1
+      in
       Pack.flush t.pack;
       Pack.sync r;
       check h1 x1 "find before filter" >>= fun () ->
@@ -430,13 +447,17 @@ module Pack = struct
       check h1 x1 "find after filter" >>= fun () ->
       let x2 = "bar" in
       let h2 = sha1 x2 in
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h2 x2;
+      let (_ : Pack.hash) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h2 x2
+      in
       Pack.flush t.pack;
       Pack.sync r;
       check h2 x2 "find before flush" >>= fun () ->
       let x3 = "toto" in
       let h3 = sha1 x3 in
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h3 x3;
+      let (_ : Pack.hash) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false w h3 x3
+      in
       Index.flush t.index;
       check h2 x2 "find after flush" >>= fun () ->
       Pack.flush t.pack;
@@ -450,7 +471,9 @@ module Pack = struct
     let* t = Context.get_pack ~lru_size:10 () in
     let v = "foo" in
     let k = sha1 v in
-    Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack k v;
+    let (_ : Pack.key) =
+      Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack k v
+    in
     Pack.flush t.pack;
     let* v1 = Pack.find t.pack k in
     Alcotest.(check (option string)) "before clear" (Some v) v1;
@@ -473,7 +496,9 @@ module Pack = struct
     let x1 = "foo" in
     let h1 = sha1 x1 in
     let find_before_and_after_sync persist file =
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h1 x1;
+      let (_ : Pack.key) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h1 x1
+      in
       persist ();
       Pack.sync r;
       Pack.clear t.pack >>= fun () ->
@@ -508,11 +533,15 @@ module Pack = struct
     let x2 = "bar" in
     let h2 = sha1 x2 in
     let find_before_and_after_sync persist file =
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h1 x1;
+      let (_ : Pack.key) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h1 x1
+      in
       persist ();
       Pack.sync r;
       Pack.clear t.pack >>= fun () ->
-      Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h2 x2;
+      let (_ : Pack.key) =
+        Pack.unsafe_append ~ensure_unique:true ~overcommit:false t.pack h2 x2
+      in
       persist ();
       let* () =
         check h1 (Some x1)