+++ CHANGES.md
@@ -143,6 +143,10 @@
   - Renamed `Irmin_mirage_git.Make` into `Irmin_mirage_git.Maker`
     (#1369, @samoht)
 
+- **irmin-pack**
+  - The `Irmin_pack.Maker` module type now no longer takes a `Conf` argument.
+    (#1641, @CraigFe)
+
 - **irmin-unix**
   - Clean up command line interface. Allow config file to be specified when
     using `Irmin_unix.Resolver.load_config` and make command line options